move functions defined in before_compile to using as overridable
The code generated in Definition includes default implementations of
functions which are supposed to be overridable. The problem is that if
the overridable implementations do pattern matching in the function head
and those patterns fail, instead of an error being raised, the default
implementation generated by Waffle is used, which hides the error and
can result in unexpected behavior (in my case, it was the ACL being
incorrectly set).

This change makes those functions overridable like the others, which
preserves the default behavior, but behaves as expected when a custom
implementation is provided.